move all routing components to new Slim\Routing aggregate folder
