--- Slim/Router.php
@@ -1,502 +1,0 @@
-<?php
-/**
- * Slim Framework (https://slimframework.com)
- *
- * @license https://github.com/slimphp/Slim/blob/4.x/LICENSE.md (MIT License)
- */
-
-declare(strict_types=1);
-
-namespace Slim;
-
-use FastRoute\RouteCollector;
-use FastRoute\RouteParser;
-use FastRoute\RouteParser\Std as StdParser;
-use InvalidArgumentException;
-use Psr\Container\ContainerInterface;
-use Psr\Http\Message\ResponseFactoryInterface;
-use Psr\Http\Message\ServerRequestInterface;
-use RuntimeException;
-use Slim\Handlers\Strategies\RequestResponse;
-use Slim\Interfaces\CallableResolverInterface;
-use Slim\Interfaces\InvocationStrategyInterface;
-use Slim\Interfaces\RouteGroupInterface;
-use Slim\Interfaces\RouteInterface;
-use Slim\Interfaces\RouterInterface;
-
-/**
- * Router
- *
- * This class organizes Slim application route objects. It is responsible
- * for registering route objects, assigning names to route objects,
- * finding routes that match the current HTTP request, and creating
- * URLs for a named route.
- */
-class Router implements RouterInterface
-{
-    /**
-     * Parser
-     *
-     * @var RouteParser
-     */
-    protected $routeParser;
-
-    /**
-     * @var CallableResolverInterface
-     */
-    protected $callableResolver;
-
-    /**
-     * @var ContainerInterface|null
-     */
-    protected $container;
-
-    /**
-     * @var InvocationStrategyInterface
-     */
-    protected $defaultInvocationStrategy;
-
-    /**
-     * Base path used in pathFor()
-     *
-     * @var string
-     */
-    protected $basePath = '';
-
-    /**
-     * Path to fast route cache file. Set to false to disable route caching
-     *
-     * @var string|null
-     */
-    protected $cacheFile = null;
-
-    /**
-     * Routes
-     *
-     * @var array
-     */
-    protected $routes = [];
-
-    /**
-     * Route counter incrementer
-     *
-     * @var int
-     */
-    protected $routeCounter = 0;
-
-    /**
-     * Route groups
-     *
-     * @var RouteGroup[]
-     */
-    protected $routeGroups = [];
-
-    /**
-     * @var Dispatcher|null
-     */
-    protected $dispatcher;
-
-    /**
-     * @var ResponseFactoryInterface
-     */
-    protected $responseFactory;
-
-    /**
-     * Create new router
-     *
-     * @param ResponseFactoryInterface      $responseFactory
-     * @param CallableResolverInterface     $callableResolver
-     * @param ContainerInterface|null       $container
-     * @param InvocationStrategyInterface   $defaultInvocationStrategy
-     * @param RouteParser                   $parser
-     */
-    public function __construct(
-        ResponseFactoryInterface $responseFactory,
-        CallableResolverInterface $callableResolver,
-        ContainerInterface $container = null,
-        InvocationStrategyInterface $defaultInvocationStrategy = null,
-        RouteParser $parser = null
-    ) {
-        $this->responseFactory = $responseFactory;
-        $this->callableResolver = $callableResolver;
-        $this->container = $container;
-        $this->defaultInvocationStrategy = $defaultInvocationStrategy ?? new RequestResponse();
-        $this->routeParser = $parser ?? new StdParser;
-    }
-
-    /**
-     * Get default route invocation strategy
-     *
-     * @return InvocationStrategyInterface
-     */
-    public function getDefaultInvocationStrategy(): InvocationStrategyInterface
-    {
-        return $this->defaultInvocationStrategy;
-    }
-
-    /**
-     * @param InvocationStrategyInterface $strategy
-     * @return self
-     */
-    public function setDefaultInvocationStrategy(InvocationStrategyInterface $strategy): RouterInterface
-    {
-        $this->defaultInvocationStrategy = $strategy;
-        return $this;
-    }
-
-    /**
-     * @return CallableResolverInterface
-     */
-    public function getCallableResolver(): CallableResolverInterface
-    {
-        return $this->callableResolver;
-    }
-
-    /**
-     * Set the base path used in pathFor()
-     *
-     * @param string $basePath
-     *
-     * @return self
-     */
-    public function setBasePath(string $basePath): self
-    {
-        $this->basePath = $basePath;
-        return $this;
-    }
-
-    /**
-     * Set path to fast route cache file. If this is false then route caching is disabled.
-     *
-     * @param string|null $cacheFile
-     * @return self
-     *
-     * @throws InvalidArgumentException
-     * @throws RuntimeException
-     */
-    public function setCacheFile(?string $cacheFile): RouterInterface
-    {
-        if ($cacheFile && file_exists($cacheFile) && !is_readable($cacheFile)) {
-            throw new RuntimeException(
-                sprintf('Router cache file `%s` is not readable', $cacheFile)
-            );
-        }
-
-        if ($cacheFile && !file_exists($cacheFile) && !is_writable(dirname($cacheFile))) {
-            throw new RuntimeException(
-                sprintf('Router cache file directory `%s` is not writable', dirname($cacheFile))
-            );
-        }
-
-        $this->cacheFile = $cacheFile;
-        return $this;
-    }
-
-    /**
-     * Add route
-     *
-     * @param  string[] $methods Array of HTTP methods
-     * @param  string   $pattern The route pattern
-     * @param  callable|string $handler The route callable
-     *
-     * @return RouteInterface
-     */
-    public function map(array $methods, string $pattern, $handler): RouteInterface
-    {
-        // Prepend parent group pattern(s)
-        if ($this->routeGroups) {
-            $pattern = $this->processGroups() . $pattern;
-        }
-
-        /** @var Route $route */
-        $route = $this->createRoute($methods, $pattern, $handler);
-        $this->routes[$route->getIdentifier()] = $route;
-        $this->routeCounter++;
-
-        return $route;
-    }
-
-    /**
-     * Dispatch router for HTTP request
-     *
-     * @param  ServerRequestInterface $request The current HTTP request object
-     *
-     * @return RoutingResults
-     */
-    public function dispatch(ServerRequestInterface $request): RoutingResults
-    {
-        $uri = '/' . ltrim(rawurldecode($request->getUri()->getPath()), '/');
-        return $this->createDispatcher()->dispatch($request->getMethod(), $uri);
-    }
-
-    /**
-     * Create a new Route object
-     *
-     * @param  string[] $methods Array of HTTP methods
-     * @param  string   $pattern The route pattern
-     * @param  callable|string $callable The route callable
-     *
-     * @return RouteInterface
-     */
-    protected function createRoute(array $methods, string $pattern, $callable): RouteInterface
-    {
-        return new Route(
-            $methods,
-            $pattern,
-            $callable,
-            $this->responseFactory,
-            $this->callableResolver,
-            $this->container,
-            $this->defaultInvocationStrategy,
-            $this->routeGroups,
-            $this->routeCounter
-        );
-    }
-
-    /**
-     * @return Dispatcher
-     */
-    protected function createDispatcher(): Dispatcher
-    {
-        if ($this->dispatcher) {
-            return $this->dispatcher;
-        }
-
-        $routeDefinitionCallback = function (RouteCollector $r) {
-            foreach ($this->getRoutes() as $route) {
-                $r->addRoute($route->getMethods(), $route->getPattern(), $route->getIdentifier());
-            }
-        };
-
-        if ($this->cacheFile) {
-            /** @var Dispatcher $dispatcher */
-            $dispatcher = \FastRoute\cachedDispatcher($routeDefinitionCallback, [
-                'dispatcher' => Dispatcher::class,
-                'routeParser' => $this->routeParser,
-                'cacheFile' => $this->cacheFile,
-            ]);
-        } else {
-            /** @var Dispatcher $dispatcher */
-            $dispatcher = \FastRoute\simpleDispatcher($routeDefinitionCallback, [
-                'dispatcher' => Dispatcher::class,
-                'routeParser' => $this->routeParser,
-            ]);
-        }
-
-        $this->dispatcher = $dispatcher;
-        return $this->dispatcher;
-    }
-
-    /**
-     * @param Dispatcher $dispatcher
-     * @return RouterInterface
-     */
-    public function setDispatcher(Dispatcher $dispatcher): RouterInterface
-    {
-        $this->dispatcher = $dispatcher;
-        return $this;
-    }
-
-    /**
-     * Get route objects
-     *
-     * @return Route[]
-     */
-    public function getRoutes(): array
-    {
-        return $this->routes;
-    }
-
-    /**
-     * Get named route object
-     *
-     * @param string $name        Route name
-     *
-     * @return RouteInterface
-     *
-     * @throws RuntimeException   If named route does not exist
-     */
-    public function getNamedRoute(string $name): RouteInterface
-    {
-        foreach ($this->routes as $route) {
-            if ($name == $route->getName()) {
-                return $route;
-            }
-        }
-        throw new RuntimeException('Named route does not exist for name: ' . $name);
-    }
-
-    /**
-     * Remove named route
-     *
-     * @param string $name        Route name
-     * @return self
-     * @throws RuntimeException   If named route does not exist
-     */
-    public function removeNamedRoute(string $name): self
-    {
-        /** @var Route $route */
-        $route = $this->getNamedRoute($name);
-        unset($this->routes[$route->getIdentifier()]);
-        return $this;
-    }
-
-    /**
-     * Process route groups
-     *
-     * @return string A group pattern to prefix routes with
-     */
-    protected function processGroups(): string
-    {
-        $pattern = "";
-        foreach ($this->routeGroups as $group) {
-            $pattern .= $group->getPattern();
-        }
-        return $pattern;
-    }
-
-    /**
-     * Add a route group to the array
-     *
-     * @param string   $pattern
-     * @param callable $callable
-     *
-     * @return RouteGroupInterface
-     */
-    public function pushGroup(string $pattern, $callable): RouteGroupInterface
-    {
-        $group = new RouteGroup($pattern, $callable, $this->callableResolver);
-        $this->routeGroups[] = $group;
-        return $group;
-    }
-
-    /**
-     * Removes the last route group from the array
-     *
-     * @return RouteGroupInterface|null The last RouteGroup, if one exists
-     */
-    public function popGroup(): ?RouteGroupInterface
-    {
-        return array_pop($this->routeGroups);
-    }
-
-    /**
-     * @param string $identifier
-     * @return RouteInterface
-     */
-    public function lookupRoute(string $identifier): RouteInterface
-    {
-        if (!isset($this->routes[$identifier])) {
-            throw new RuntimeException('Route not found, looks like your route cache is stale.');
-        }
-        return $this->routes[$identifier];
-    }
-
-    /**
-     * Build the path for a named route excluding the base path
-     *
-     * @param string $name        Route name
-     * @param array  $data        Named argument replacement data
-     * @param array  $queryParams Optional query string parameters
-     *
-     * @return string
-     *
-     * @throws RuntimeException         If named route does not exist
-     * @throws InvalidArgumentException If required data not provided
-     */
-    public function relativePathFor(string $name, array $data = [], array $queryParams = []): string
-    {
-        $route = $this->getNamedRoute($name);
-        $pattern = $route->getPattern();
-
-        $routeDatas = $this->routeParser->parse($pattern);
-        // $routeDatas is an array of all possible routes that can be made. There is
-        // one routedata for each optional parameter plus one for no optional parameters.
-        //
-        // The most specific is last, so we look for that first.
-        $routeDatas = array_reverse($routeDatas);
-
-        $segments = [];
-        $segmentName = '';
-        foreach ($routeDatas as $routeData) {
-            foreach ($routeData as $item) {
-                if (is_string($item)) {
-                    // this segment is a static string
-                    $segments[] = $item;
-                    continue;
-                }
-
-                // This segment has a parameter: first element is the name
-                if (!array_key_exists($item[0], $data)) {
-                    // we don't have a data element for this segment: cancel
-                    // testing this routeData item, so that we can try a less
-                    // specific routeData item.
-                    $segments = [];
-                    $segmentName = $item[0];
-                    break;
-                }
-                $segments[] = $data[$item[0]];
-            }
-            if (!empty($segments)) {
-                // we found all the parameters for this route data, no need to check
-                // less specific ones
-                break;
-            }
-        }
-
-        if (empty($segments)) {
-            throw new InvalidArgumentException('Missing data for URL segment: ' . $segmentName);
-        }
-        $url = implode('', $segments);
-
-        if ($queryParams) {
-            $url .= '?' . http_build_query($queryParams);
-        }
-
-        return $url;
-    }
-
-
-    /**
-     * Build the path for a named route including the base path
-     *
-     * @param string $name        Route name
-     * @param array  $data        Named argument replacement data
-     * @param array  $queryParams Optional query string parameters
-     *
-     * @return string
-     *
-     * @throws RuntimeException         If named route does not exist
-     * @throws InvalidArgumentException If required data not provided
-     */
-    public function pathFor(string $name, array $data = [], array $queryParams = []): string
-    {
-        $url = $this->relativePathFor($name, $data, $queryParams);
-
-        if ($this->basePath) {
-            $url = $this->basePath . $url;
-        }
-
-        return $url;
-    }
-
-    /**
-     * Build the path for a named route.
-     *
-     * This method is deprecated. Use pathFor() from now on.
-     *
-     * @param string $name        Route name
-     * @param array  $data        Named argument replacement data
-     * @param array  $queryParams Optional query string parameters
-     *
-     * @return string
-     *
-     * @throws RuntimeException         If named route does not exist
-     * @throws InvalidArgumentException If required data not provided
-     */
-    public function urlFor(string $name, array $data = [], array $queryParams = []): string
-    {
-        trigger_error('urlFor() is deprecated. Use pathFor() instead.', E_USER_DEPRECATED);
-        return $this->pathFor($name, $data, $queryParams);
-    }
-}
